[feat] new variable in dataset kwargs: create_link

+++ lmms_eval/api/task.py
@@ -938,6 +938,7 @@ class ConfigurableTask(Task):
                     force_download = dataset_kwargs.get("force_download", False)
                     force_unzip = dataset_kwargs.get("force_unzip", False)
                     revision = dataset_kwargs.get("revision", "main")
+                    create_link = dataset_kwargs.get("create_link", False)
                     cache_path = snapshot_download(repo_id=self.DATASET_PATH, revision=revision, repo_type="dataset", force_download=force_download, etag_timeout=60)
                     zip_files = glob(os.path.join(cache_path, "**/*.zip"), recursive=True)
                     tar_files = glob(os.path.join(cache_path, "**/*.tar*"), recursive=True)
@@ -1001,14 +1002,15 @@ class ConfigurableTask(Task):
                             if not os.path.exists(os.path.join(cache_dir, os.path.basename(base_name))):
                                 untar_video_data(output_tar)
 
-                    # Link cache_path to cache_dir.
-                    if not os.path.exists(cache_dir) or os.path.islink(cache_dir):
-                        if os.path.islink(cache_dir):
-                            os.remove(cache_dir)
-                            print(f"Removed existing symbolic link: {cache_dir}")
-                        # Create a new symbolic link
-                        os.symlink(cache_path, cache_dir)
-                        print(f"Symbolic link created successfully: {cache_path} -> {cache_dir}")
+                    # Link cache_path to cache_dir if needed.
+                    if create_link:
+                        if not os.path.exists(cache_dir) or os.path.islink(cache_dir):
+                            if os.path.islink(cache_dir):
+                                os.remove(cache_dir)
+                                eval_logger.info(f"Removed existing symbolic link: {cache_dir}")
+                            # Create a new symbolic link
+                            os.symlink(cache_path, cache_dir)
+                            eval_logger.info(f"Symbolic link created successfully: {cache_path} -> {cache_dir}")
 
                 accelerator.wait_for_everyone()
                 dataset_kwargs.pop("cache_dir")
@@ -1027,6 +1029,9 @@ class ConfigurableTask(Task):
 
             if "local_files_only" in dataset_kwargs:
                 dataset_kwargs.pop("local_files_only")
+
+            if "create_link" in dataset_kwargs:
+                dataset_kwargs.pop("create_link")
 
         self.dataset = datasets.load_dataset(
             path=self.DATASET_PATH,
